requirements.txt
torch==1.13.1
torchvision==0.14.1
opencv-python
Pillow
PyYAML
numpy
urllib3<2
